Fix richtext grammar
Issue was that `closed` being a [text | {…}]* didn't work
because one of the alternative was `text`.

What happens with two `text`s that follow each other? It's impossible to
tell when the first end and the second starts.

So we replaced [text | {…}]* by alternating text and {…}. This seems to
work much better.