Add trait-driven metadata parsing.
Refactor the parsing module to separate "interpreting" the data from
parsing it. (IMO this probably should be reverted to avoid collecting
twice into a Vec, and improve error messages, but that's for the
future).

Now the parser uses an "interpreting context" to turn the parse tree (a
single `Vec` containing a bunch of `&str`) into `RichText` sections and
collections of `ModifyBox`es.