Add dependency-driven updating BROKEN
Completely rework RichText.

- Now store modifier + span instead of duplicating modifiers in
  individual sections
- Keep track of dependencies of modifiers, now only modifiers that has an
  updated dependency will run.
- Use a string interner to update binding names, now binding
  manipulation requires no strings. Especially useful for trackers.

Remove "by type" bindings, we cannot support it (yet) with our new
setup. And I doubt it will be useful in the long term.

The interner requires access to the world when making the RichText, so
we can't create RichText in a system, we instead create a MakeRichtext,
which a system picks up and consumes. It the adds the RichTextData to
the entity in question.

The examples still do not run, but all tests pass.