Change RichText representation
This supports:

- Nested modify
- modify dependencies (m2m deps)
- Fine-grained field-based modify dependency (c2m deps)
- Culling static modifiers from RichText

The new .md files do a bit of explanation on how the implementation
works. But to give a short idea here: We build a directed acyclic graph
with:

- C inputs (nºfields our Modify can touch)
- S*C ouputs (S is nºsections)
- M nodes (nºdynamic Modify)

Each node is stored in `RichText::modifies` (this field name is a
typo and will be changed).

`RichText`'s other fields store relationships between inputs, nodes and
outputs.

- `direct_deps`: c2m, list of Modify to trigger when an input changes
- `modify_deps`: m2m, list of Modify to trigger when a Modify is ran
- `dynamic`: b2m, list of Modify to trigger for a binding when its value
  changes
- `binding_masks`: b2s, list of sections NOT to update when running a
  Modify tied to a specific binding
- `root_mask`: c2s, list of sections NOT to update when running a Modify
  tied to a specific input change.

Note that `dynamic` should really not point to a `Modify`,
since it will be user-provied.

Each of those fields use a different specialized data structure, all
defined in `store` module.