Add handling of pass-by-reference for modify enum
Now `foo: &Foo` is added as a field of type `Foo` to the modify enum
variant corresponding to the modify function accepting
`modify_fn(foo: &Foo)` and is passed by reference.

This allows using non-copy types as modify enum variant field.

It is smart enough to not do it with `&'static` types (actually any type
with explicit lifetime).