Refactor resolver + b2m
Make the `modify` field of `Modifier` optional, to handle binding
modifiers. Instead of having a dedicated buffer for the binding ranges,
we store the range with other modifiers in the `modifiers` buffer.

We replace the `bindings` range buffer by `b2m` a list of binding to
index in `modifiers` to read to learn of range.