Fix resolver using the wrong range for bindings
In commit 3c3d030fb, the change to `binding_range` to reduce the search
space for binding was incorrect.

More precisely, the index in the subset slice isn't the same as the
index in the whole array. This means that when later we get
the value associated with the key we searched using search_by_key,
`self.b2m[index]`, we aren't getting the correct value!

The solution is either to get the value from `subset` or add back
`start_at` to the index. I opted for the second solution, since we do
use `index` later.