Fix modifier masks
The "root mask" was wrong most of the time, we now build an individual
mask of sections not-to-update per modifier.

this new `masks` field only contain vaues when it matters, so it doesn't
bloat memory usage like at all. Memory usage is 4 bytes per modifier
mostly.

When updating sections, we check against the modfier's mask if we should
skip any.